# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Base go API.",
        "title": "Base Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1234",
    "paths": {
        "/cmd/execute-script": {
            "post": {
                "description": "Execute command line in a text script file",
                "tags": [
                    "Command"
                ],
                "summary": "Execute Script",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incoming.ExcecuteScriptParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ExecuteScriptOutgoing"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelBadRequestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelInternalErr"
                        }
                    }
                }
            }
        },
        "/cmd/kill-process": {
            "post": {
                "description": "Kill a process is runing",
                "tags": [
                    "Command"
                ],
                "summary": "Kill Process",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incoming.KillProcessParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outgoing.KillProcessOutgoing"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelBadRequestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelInternalErr"
                        }
                    }
                }
            }
        },
        "/cmd/list-process": {
            "get": {
                "description": "Get all process runing in computer",
                "tags": [
                    "Command"
                ],
                "summary": "List Process",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ListProcessOutgoing"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelInternalErr"
                        }
                    }
                }
            }
        },
        "/cmd/open-file": {
            "post": {
                "description": "Find and open file by path",
                "tags": [
                    "Command"
                ],
                "summary": "Find File",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incoming.GetFileParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outgoing.OpenFileSucessOutgoing"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelBadRequestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outgoing.OpenFileFailedOutgoing"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Login to app",
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outgoing.LoginOutgoing"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelBadRequestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelInternalErr"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register new account",
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outgoing.RegisterOutgoing"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelBadRequestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outgoing.ModelInternalErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "incoming.ExcecuteScriptParams": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string",
                    "example": ".\\script.ps1"
                }
            }
        },
        "incoming.GetFileParams": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string",
                    "example": ".\\script.ps1"
                }
            }
        },
        "incoming.KillProcessParams": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "notepad.exe"
                },
                "pId": {
                    "type": "string",
                    "example": "12345"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "456"
                },
                "password": {
                    "type": "string",
                    "example": "1234"
                },
                "token": {
                    "type": "string",
                    "example": ""
                },
                "userName": {
                    "type": "string",
                    "example": "truong"
                }
            }
        },
        "outgoing.ExecuteScriptOutgoing": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "excute done!"
                }
            }
        },
        "outgoing.KillProcessOutgoing": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "process killed!"
                }
            }
        },
        "outgoing.ListProcessOutgoing": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "notepad.exe      1021"
                }
            }
        },
        "outgoing.LoginOutgoing": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "outgoing.ModelBadRequestErr": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid query params"
                }
            }
        },
        "outgoing.ModelInternalErr": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Internal sever error"
                }
            }
        },
        "outgoing.OpenFileFailedOutgoing": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "cannot find file with path"
                }
            }
        },
        "outgoing.OpenFileSucessOutgoing": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "opening file"
                }
            }
        },
        "outgoing.RegisterOutgoing": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.User"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}